Add platform selector
Uses a hook to pull the current project selection and returns no project
if it's not cross platform. Also returns the platform selector
component. This refactors a bunch of files to remove project from props
and just calls the hook
