+++ static/app/views/performance/mobile/screenload/index.tsx
@@ -1,4 +1,3 @@
-import {useMemo} from 'react';
 import styled from '@emotion/styled';
 
 import {Breadcrumbs} from 'sentry/components/breadcrumbs';
@@ -15,16 +14,13 @@ import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
 import {PageAlert, PageAlertProvider} from 'sentry/utils/performance/contexts/pageAlert';
 import useOrganization from 'sentry/utils/useOrganization';
-import usePageFilters from 'sentry/utils/usePageFilters';
-import useProjects from 'sentry/utils/useProjects';
 import {useOnboardingProject} from 'sentry/views/performance/browser/webVitals/utils/useOnboardingProject';
 import {ScreensView, YAxis} from 'sentry/views/performance/mobile/screenload/screens';
-import {PlatformSelector} from 'sentry/views/performance/mobile/screenload/screens/platformSelector';
-import {isCrossPlatform} from 'sentry/views/performance/mobile/screenload/screens/utils';
 import {
   MODULE_DESCRIPTION,
   MODULE_DOC_LINK,
 } from 'sentry/views/performance/mobile/screenload/settings';
+import usePlatformSelector from 'sentry/views/performance/mobile/usePlatformSelector';
 import {ModulePageProviders} from 'sentry/views/performance/modulePageProviders';
 import Onboarding from 'sentry/views/performance/onboarding';
 import {useModuleBreadcrumbs} from 'sentry/views/performance/utils/useModuleBreadcrumbs';
@@ -33,15 +29,7 @@ import {ReleaseComparisonSelector} from 'sentry/views/starfish/components/releas
 export function PageloadModule() {
   const organization = useOrganization();
   const onboardingProject = useOnboardingProject();
-  const {selection} = usePageFilters();
-  const {projects} = useProjects();
-
-  const project = useMemo(() => {
-    if (selection.projects.length !== 1) {
-      return null;
-    }
-    return projects.find(p => p.id === String(selection.projects));
-  }, [projects, selection.projects]);
+  const {isProjectCrossPlatform, PlatformSelector} = usePlatformSelector();
 
   const crumbs = useModuleBreadcrumbs('screen_load');
 
@@ -63,7 +51,7 @@ export function PageloadModule() {
           </Layout.HeaderContent>
           <Layout.HeaderActions>
             <ButtonBar gap={1}>
-              {project && isCrossPlatform(project) && <PlatformSelector />}
+              {isProjectCrossPlatform && <PlatformSelector />}
               <FeedbackWidgetButton />
             </ButtonBar>
           </Layout.HeaderActions>
@@ -87,11 +75,7 @@ export function PageloadModule() {
                 </OnboardingContainer>
               )}
               {!onboardingProject && (
-                <ScreensView
-                  yAxes={[YAxis.TTID, YAxis.TTFD]}
-                  chartHeight={240}
-                  project={project}
-                />
+                <ScreensView yAxes={[YAxis.TTID, YAxis.TTFD]} chartHeight={240} />
               )}
             </ErrorBoundary>
           </Layout.Main>
