+++ static/app/views/performance/mobile/screenload/screenLoadSpans/table.tsx
@@ -13,7 +13,6 @@ import Pagination from 'sentry/components/pagination';
 import {Tooltip} from 'sentry/components/tooltip';
 import {t, tct} from 'sentry/locale';
 import type {NewQuery} from 'sentry/types/organization';
-import type {Project} from 'sentry/types/project';
 import {browserHistory} from 'sentry/utils/browserHistory';
 import type {TableDataRow} from 'sentry/utils/discover/discoverQuery';
 import type {MetaType} from 'sentry/utils/discover/eventView';
@@ -38,8 +37,8 @@ import {
   PLATFORM_QUERY_PARAM,
 } from 'sentry/views/performance/mobile/screenload/screens/platformSelector';
 import {useTableQuery} from 'sentry/views/performance/mobile/screenload/screens/screensTable';
-import {isCrossPlatform} from 'sentry/views/performance/mobile/screenload/screens/utils';
 import {MODULE_DOC_LINK} from 'sentry/views/performance/mobile/screenload/settings';
+import usePlatformSelector from 'sentry/views/performance/mobile/usePlatformSelector';
 import {useModuleURL} from 'sentry/views/performance/utils/useModuleURL';
 import {
   PRIMARY_RELEASE_ALIAS,
@@ -57,7 +56,6 @@ const {SPAN_SELF_TIME, SPAN_DESCRIPTION, SPAN_GROUP, SPAN_OP, PROJECT_ID} =
 
 type Props = {
   primaryRelease?: string;
-  project?: Project | null;
   secondaryRelease?: string;
   transaction?: string;
 };
@@ -66,20 +64,19 @@ export function ScreenLoadSpansTable({
   transaction,
   primaryRelease,
   secondaryRelease,
-  project,
 }: Props) {
   const moduleURL = useModuleURL('screen_load');
   const location = useLocation();
   const {selection} = usePageFilters();
   const organization = useOrganization();
   const cursor = decodeScalar(location.query?.[MobileCursors.SPANS_TABLE]);
+  const {isProjectCrossPlatform} = usePlatformSelector();
 
   const spanOp = decodeScalar(location.query[SpanMetricsField.SPAN_OP]) ?? '';
   const {hasTTFD, isLoading: hasTTFDLoading} = useTTFDConfigured([
     `transaction:"${transaction}"`,
   ]);
 
-  const hasPlatformSelectFeature = organization.features.includes('spans-first-ui');
   const platform =
     decodeScalar(location.query[PLATFORM_QUERY_PARAM]) ??
     localStorage.getItem(PLATFORM_LOCAL_STORAGE_KEY) ??
@@ -95,16 +92,15 @@ export function ScreenLoadSpansTable({
         : [`span.op:[${TTID_CONTRIBUTING_SPAN_OPS.join(',')}]`]),
     ]);
 
-    if (project && isCrossPlatform(project) && hasPlatformSelectFeature) {
+    if (isProjectCrossPlatform) {
       searchQuery.addFilterValue('os.name', platform);
     }
 
     return appendReleaseFilters(searchQuery, primaryRelease, secondaryRelease);
   }, [
-    hasPlatformSelectFeature,
+    isProjectCrossPlatform,
     platform,
     primaryRelease,
-    project,
     secondaryRelease,
     spanOp,
     transaction,
