+++ static/app/views/performance/mobile/screenload/screens/screensTable.tsx
@@ -10,7 +10,6 @@ import type {CursorHandler} from 'sentry/components/pagination';
 import Pagination from 'sentry/components/pagination';
 import {Tooltip} from 'sentry/components/tooltip';
 import {t, tct} from 'sentry/locale';
-import type {Project} from 'sentry/types/project';
 import type {TableData, TableDataRow} from 'sentry/utils/discover/discoverQuery';
 import {useDiscoverQuery} from 'sentry/utils/discover/discoverQuery';
 import type {MetaType} from 'sentry/utils/discover/eventView';
@@ -24,6 +23,7 @@ import usePageFilters from 'sentry/utils/usePageFilters';
 import TopResultsIndicator from 'sentry/views/discover/table/topResultsIndicator';
 import type {TableColumn} from 'sentry/views/discover/table/types';
 import {TOP_SCREENS} from 'sentry/views/performance/mobile/constants';
+import usePlatformSelector from 'sentry/views/performance/mobile/usePlatformSelector';
 import {useModuleURL} from 'sentry/views/performance/utils/useModuleURL';
 import {
   PRIMARY_RELEASE_ALIAS,
@@ -38,21 +38,14 @@ type Props = {
   isLoading: boolean;
   pageLinks: string | undefined;
   onCursor?: CursorHandler;
-  project?: Project | null;
 };
 
-export function ScreensTable({
-  data,
-  eventView,
-  isLoading,
-  pageLinks,
-  onCursor,
-  project,
-}: Props) {
+export function ScreensTable({data, eventView, isLoading, pageLinks, onCursor}: Props) {
   const moduleURL = useModuleURL('screen_load');
   const location = useLocation();
   const organization = useOrganization();
   const {primaryRelease, secondaryRelease} = useReleaseSelection();
+  const {project} = usePlatformSelector();
   const eventViewColumns = eventView.getColumns();
 
   const columnNameMap = {
